tests/scripts/requirements.txt
kafka-python==1.3.5
